requirements.txt
SQLAlchemy
scrapy_splash
scrapy
slpp-23